refactor: renames storage/listing to storage/sell
