+++ src/components/pages/storage/sell/StorageSellDone.tsx
@@ -2,10 +2,10 @@ import React from 'react'
 import TxCompletePageTemplate from 'components/templates/TxCompletePageTemplate'
 import { JobDoneBox } from 'components/molecules'
 
-const StorageOfferListed = () => (
+const StorageSellDone = () => (
   <TxCompletePageTemplate>
     <JobDoneBox text="Your offer has been listed." />
   </TxCompletePageTemplate>
 )
 
-export default StorageOfferListed
+export default StorageSellDone
